тесты для FilmCard в FilmCard.e2e.test
