+++ src/components/film-card/film-card.e2e.test.js
@@ -14,17 +14,49 @@ Enzyme.configure({
   adapter: new Adapter(),
 });
 
-it(`Should be hovered`, () => {
-  const onCardHover = jest.fn();
+describe(`Should VideoPlayer play and pause`, () => {
+  it(`Should be hovered`, () => {
+    const onCardHover = jest.fn();
 
-  const filmCard = shallow(
-      <FilmCard
-        filmInfo={movie}
-        onTitleClick={()=>{}}
-        onCardHover={onCardHover}
-      />
-  );
-  filmCard.props().onMouseOver();
+    const filmCard = shallow(
+        <FilmCard
+          filmInfo={movie}
+          onTitleClick={()=>{}}
+          onCardHover={onCardHover}
+        />
+    );
+    filmCard.props().onMouseOver();
 
-  expect(onCardHover.mock.calls.length).toBe(1);
+    expect(onCardHover.mock.calls.length).toBe(1);
+  });
+
+  it(`Should VideoPlayer pause`, () => {
+    const filmCard = shallow(
+        <FilmCard
+          filmInfo={movie}
+          onTitleClick={()=>{}}
+          onCardHover={()=>{}}
+        />
+    );
+
+    const movieCard = filmCard.find(`.small-movie-card`);
+    movieCard.simulate(`mouseLeave`);
+
+    expect(filmCard.state(`isVideoPlaying`)).toEqual(false);
+  });
+
+  it(`Should VideoPlayer play`, () => {
+    const filmCard = shallow(
+        <FilmCard
+          filmInfo={movie}
+          onTitleClick={()=>{}}
+          onCardHover={()=>{}}
+        />
+    );
+
+    const movieCard = filmCard.find(`.small-movie-card`);
+    movieCard.simulate(`mouseEnter`);
+
+    expect(filmCard.state(`isVideoPlaying`)).toEqual(true);
+  });
 });
